+++ lib/src/enum/enum_device_type.dart
@@ -1,7 +1,0 @@
-enum EnumDeviceType {
-  phone,
-  tablet,
-  pc,
-  // レイアウトがデバイスの種類を問わない場合（レスポンシブデザイン等）の指定
-  any
-}