+++ lib/src/enum/enum_window_class.dart
@@ -1,3 +1,4 @@
 /// I created this enum with reference to:
 /// https://m3.material.io/foundations/layout/applying-layout/window-size-classes
-enum EnumWindowClass { compact, medium, expanded }
+/// The any class was added for convenience when writing layout files.
+enum EnumWindowClass { compact, medium, expanded, any }