+++ lib/src/enum/spwml_enum_all.dart
@@ -1,6 +1,5 @@
 export 'enum_btn_type.dart';
 export 'enum_clip_type.dart';
-export 'enum_device_type.dart';
 export 'enum_font_style.dart';
 export 'enum_font_weight.dart';
 export 'enum_img_type.dart';
@@ -12,3 +11,4 @@ export 'enum_switch_btn_type.dart';
 export 'enum_text_decoration.dart';
 export 'enum_text_decoration_style.dart';
 export 'enum_textfield_params.dart';
+export 'enum_window_class.dart';